refactor(fusion): the stream tests are a file of their own
`multi.rs` was 2288 lines, 1855 of them tests. They move to
`multi/tests.rs`, a child module, so they keep reaching the private
state they drive. The suite's doc goes with them, as the file's own.
